+++ File_Uploads/forms.py
@@ -1,0 +1,6 @@
+from django.forms import ModelForm
+from File_Uploads.models import FileUpload
+class UploadFileForm(ModelForm):
+    class Meta:
+        model = FileUpload
+        fields = ['file']
