+++ Induction/settings.py
@@ -44,7 +44,7 @@ INSTALLED_APPS = [
     'Models',
     'Rest_framework_status',
     'Csrf',
-
+    'File_Uploads'
 ]
 
 MIDDLEWARE = [
